--- apps/client/mutations/restaurant/index.ts
@@ -1,4 +1,0 @@
-export { default as usePostScrap } from "./usePostScrap";
-export { default as useDeleteScrap } from "./useDeleteScrap";
-export { default as usePostLike } from "./usePostLike";
-export { default as useDeleteLike } from "./useDeleteLike";
