--- apps/client/queries/restaurant/index.ts
@@ -1,4 +1,0 @@
-export { default as Keys } from "./keys";
-export { default as useFetchRestaurant } from "./useFetchRestaurant";
-export { default as useFetchLikePercent } from "./useFetchLikePercent";
-export { default as useFetchReviews } from "./useFetchReviews";
