--- apps/client/queries/restaurant/keys.ts
@@ -1,7 +1,0 @@
-const Keys = {
-  restaurant: (id: number) => ["restaurant", id],
-  reviews: (id: number) => ["reviews", id],
-  likePercent: (id: number) => ["likePercent", id],
-};
-
-export default Keys;
