--- apps/client/queries/restaurant/useFetchLikePercent.ts
@@ -1,9 +1,0 @@
-import { getLikePercent } from "~/apis/restaurant";
-import { useSuspendedQuery } from "~/hooks";
-import Keys from "./keys";
-
-const useFetchLikePercent = (id: number) => {
-  return useSuspendedQuery(Keys.likePercent(id), () => getLikePercent({ id }));
-};
-
-export default useFetchLikePercent;
