--- apps/client/queries/restaurant/useFetchReviews.ts
@@ -1,9 +1,0 @@
-import { getReviews } from "~/apis/restaurant";
-import { useSuspendedQuery } from "~/hooks";
-import Keys from "./keys";
-
-const useFetchReviews = (id: number) => {
-  return useSuspendedQuery(Keys.reviews(id), () => getReviews({ id }));
-};
-
-export default useFetchReviews;
