+++ apps/client/server/restaurant/index.ts
@@ -1,0 +1,4 @@
+export * from "./queries";
+export * from "./mutations";
+export * from "./apis";
+export { default as Keys } from "./keys";
